Borgia_backend/Borgia/requirements/common.txt
Django==2.1.11
django-bootstrap-form==3.4
django-static-precompiler==1.8.2
openpyxl==2.5.12
Pillow
django-filter
djangorestframework
django-utils-six